final ver. before 1 evaluation

+++ ft_calloc.c
@@ -6,13 +6,13 @@
 /*   By: mraymun <mraymun@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/10/31 18:11:13 by mraymun           #+#    #+#             */
-/*   Updated: 2020/10/31 19:12:15 by mraymun          ###   ########.fr       */
+/*   Updated: 2020/11/04 18:56:34 by mraymun          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	*ft_calloc(int count, int size)
+void	*ft_calloc(size_t count, size_t size)
 {
 	int		bytes;
 	void	*ret;
