+++ ft_strncmp.c
@@ -3,16 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strncmp.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mraymun <mraymun@student.21-school.ru>     +#+  +:+       +#+        */
+/*   By: mraymun <mraymun@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/10/28 17:08:49 by mraymun           #+#    #+#             */
-/*   Updated: 2020/10/31 12:23:56 by mraymun          ###   ########.fr       */
+/*   Updated: 2020/11/04 18:55:29 by mraymun          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int		ft_strncmp(char *s1, char *s2, int n)
+int		ft_strncmp(char *s1, char *s2, size_t n)
 {
 	unsigned char	u1;
 	unsigned char	u2;
