+++ ft_strnlen.c
@@ -6,15 +6,15 @@
 /*   By: mraymun <mraymun@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/10/31 18:23:52 by mraymun           #+#    #+#             */
-/*   Updated: 2020/10/31 18:24:18 by mraymun          ###   ########.fr       */
+/*   Updated: 2020/11/04 18:58:40 by mraymun          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strnlen(char *str, int maxlen)
+size_t	ft_strnlen(char *str, size_t maxlen)
 {
-	int		len;
+	size_t	len;
 
 	len = 0;
 	while (len < maxlen)
